add invoices pdf, sale report csv, discount, tax, and total calculations and

+++ backend/src/services/vendorService.js
@@ -1,4 +1,5 @@
 const Vendor = require('../models/vendorModel');
+const GoodsPurchase = require('../models/goodsPurchaseModel');
 const ApiError = require('../utils/ApiError');
 const journalService = require('./journalService');
 const { ACCOUNT } = require('../utils/finance');
@@ -71,6 +72,9 @@ async function deleteVendor(id) {
   const vendor = await getVendorById(id);
   if (vendor.outstanding > 0) {
     throw ApiError.badRequest('Vendor has an outstanding balance and cannot be deleted');
+  }
+  if (await GoodsPurchase.exists({ vendor: id })) {
+    throw ApiError.badRequest('Vendor has purchase history and cannot be deleted');
   }
   await vendor.deleteOne();
 }
